fix: ST_CoordDim failing with wrong return type (#15203)
Summary:
ST_CoordDim return type should be tinyint based on : https://github.com/prestodb/presto/issues/26344

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15203

Reviewed By: bikramSingh91

Differential Revision: D85060868

Pulled By: pedroerp

fbshipit-source-id: f8d61d597dc2dde359b133d0a4a605b93c27cf4f

+++ velox/functions/prestosql/tests/GeometryFunctionsTest.cpp
@@ -2171,9 +2171,9 @@ TEST_F(GeometryFunctionsTest, testStConvexHull) {
 
 TEST_F(GeometryFunctionsTest, testStCoordDim) {
   const auto testStCoordDimFunc = [&](const std::optional<std::string>& wkt,
-                                      const std::optional<int32_t>& expected) {
-    std::optional<int32_t> result =
-        evaluateOnce<int32_t>("ST_CoordDim(ST_GeometryFromText(c0))", wkt);
+                                      const std::optional<int8_t>& expected) {
+    std::optional<int8_t> result =
+        evaluateOnce<int8_t>("ST_CoordDim(ST_GeometryFromText(c0))", wkt);
 
     if (expected.has_value()) {
       ASSERT_TRUE(result.has_value());
